Refactoring
Since Spring Security has been added to the project
CommentRestControllerTest and SuggestionRestControllerTest
fail.
With the changes of this commit Spring Security is not used for their
unit tests.

Added two unit tests to CommentServiceTest to check the correctness of
the method "getCommentsBySuggestionId" added to CommentService in the
merge between "persistence and services" and "controllers"