Updated SuggestionsWebAppE2E
removed the first test that was only an initial test and created two
tests that simulate the behavior of a generic user and the behavior of
an admin.
Updated the pom, added the properties needed by the authentication
mechanism: these properties are used during the E2E test by the
webdriver and by the application run on the tomcat server.
Updated also the application.properties in order to be able to run the
E2E tests locally and not only in the GitHub actions execution.
.classpath updated by Eclipse
Added logback-test.xml to limit the log level of packages
used by the WebDriverManager when it downloads the driver (otherwise it
log too many lines and in the CI this might exhaust the console buffer
available).